chore: fix debugger on windows
