+++ package.json
@@ -43,7 +43,6 @@
     "@types/jest": "29.1.0",
     "@types/node": "16",
     "@types/rimraf": "3.0.2",
-    "chalk": "^5.3.0",
     "commander": "10.0.1",
     "cross-env": "^7.0.3",
     "husky": "^7.0.4",
