+++ scripts/package.json
@@ -11,11 +11,14 @@
     "typecheck": "pnpm tsc --noEmit"
   },
   "dependencies": {
+    "@actions/core": "1.10.0",
     "@iarna/toml": "2.2.5",
     "@pnpm/find-workspace-packages": "6.0.8",
     "@pnpm/logger": "5.0.0",
     "node-fetch": "2.6.7",
-    "zx": "^7.2.1", 
-    "@actions/core": "1.10.0"
+    "zx": "^7.2.1"
+  },
+  "devDependencies": {
+    "chalk": "^5.3.0"
   }
 }
